backend/requirement.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-dotenv
requests
supabase
